refactor: Rename `…kage/compile/events/api-gateway/lib/usagePlanKeys.js`
