+++ lib/plugins/aws/package/compile/events/api-gateway/index.js
@@ -10,7 +10,7 @@ const compileRestApi = require('./lib/rest-api');
 const compileRequestValidators = require('./lib/request-validator');
 const compileApiKeys = require('./lib/api-keys');
 const compileUsagePlan = require('./lib/usage-plan');
-const compileUsagePlanKeys = require('./lib/usagePlanKeys');
+const compileUsagePlanKeys = require('./lib/usage-plan-keys');
 const compileResources = require('./lib/resources');
 const compileCors = require('./lib/cors');
 const compileMethods = require('./lib/method/index');
